requirements.txt
beautifulsoup4==4.12.3
requests==2.32.3
pytz==2024.1
PyYAML==6.0.2